xai api tests into cli directory
